requirements.txt
# 必須ライブラリ
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0
tiktoken>=0.5.0
pandas>=2.0.0
python-dotenv>=1.0.0

# オプションライブラリ
matplotlib>=3.7.0
rich>=13.0.0
pydantic>=2.0.0
colorama>=0.4.6
streamlit>=1.28.0 